+++ src/diffusers/schedulers/scheduling_lcm.py
@@ -244,7 +244,7 @@ class LCMScheduler(SchedulerMixin, ConfigMixin):
 
         # setable values
         self.num_inference_steps = None
-        self.timesteps = torch.from_numpy(np.arange(0, num_train_timesteps)[::-1].copy().astype(np.int64))
+        self.timesteps = torch.from_numpy(np.arange(0, num_train_timesteps)[::-1].copy().astype(np.int32))
         self.custom_timesteps = False
 
         self._step_index = None
@@ -441,7 +441,7 @@ class LCMScheduler(SchedulerMixin, ConfigMixin):
                     f" unexpected results when using this timestep schedule."
                 )
 
-            timesteps = np.array(timesteps, dtype=np.int64)
+            timesteps = np.array(timesteps, dtype=np.int32)
             self.num_inference_steps = len(timesteps)
             self.custom_timesteps = True
 
@@ -479,7 +479,7 @@ class LCMScheduler(SchedulerMixin, ConfigMixin):
             lcm_origin_timesteps = lcm_origin_timesteps[::-1].copy()
             # Select (approximately) evenly spaced indices from lcm_origin_timesteps.
             inference_indices = np.linspace(0, len(lcm_origin_timesteps), num=num_inference_steps, endpoint=False)
-            inference_indices = np.floor(inference_indices).astype(np.int64)
+            inference_indices = np.floor(inference_indices).astype(np.int32)
             timesteps = lcm_origin_timesteps[inference_indices]
 
         self.timesteps = torch.from_numpy(timesteps).to(device=device, dtype=torch.long)
